+++ .eslintignore
@@ -1,4 +1,4 @@
 node_modules
 
 app.js
-test/
+tests/
