+++ .eslintrc
@@ -32,6 +32,8 @@
         "consistent-return": 0
     },
     "globals": {
-        "Promise": true
+        "Promise": true,
+        "describe": true,
+        "it": true
     }
 }
