+++ package.json
@@ -1,31 +1,32 @@
 {
-    "name": "6sense",
-    "version": "0.1.4",
-    "description": "Counting wifi searching devices",
-    "main": "./src/index.js",
-    "scripts": {
-        "postinstall": "if [ -d .git ]; then cp ./git-hooks/pre-commit ./.git/hooks/pre-commit; chmod a+x ./.git/hooks/pre-commit; fi",
-        "lint": "eslint ./",
-        "test": "mocha test/*"
-    },
-    "author": "ants <contact@ants.builders> (http://ants.builders)",
-    "license": "MIT",
-    "dependencies": {
-        "csv-parser": "^1.5.2",
-        "es6-shim": "^0.31.0",
-        "fast-csv": "^0.6.0",
-        "machina": "^1.0.0",
-        "moment": "^2.10.3",
-        "moment-timezone": "^0.3.1",
-        "node-watch": "^0.3.4",
-        "protocol-buffers": "^3.0.0",
-        "split": "^0.3.3",
-        "through": "^2.3.7"
-    },
-    "devDependencies": {
-        "chai": "^2.3.0",
-        "chai-as-promised": "^5.0.0",
-        "mocha": "^2.2.5",
-        "eslint": "^0.21.0"
-    }
+  "name": "6sense",
+  "version": "0.2.0",
+  "description": "Counting wifi/bluetooth devices",
+  "main": "./src/index.js",
+  "scripts": {
+    "postinstall": "if [ -d .git ]; then cp ./git-hooks/pre-commit ./.git/hooks/pre-commit; chmod a+x ./.git/hooks/pre-commit; fi",
+    "lint": "eslint ./",
+    "test": "mocha tests/*"
+  },
+  "author": "ants <contact@ants.builders> (http://ants.builders)",
+  "license": "MIT",
+  "dependencies": {
+    "csv-parser": "^1.7.0",
+    "es6-shim": "^0.31.0",
+    "fast-csv": "^0.6.0",
+    "machina": "^1.1.2",
+    "moment": "^2.10.3",
+    "moment-timezone": "^0.3.1",
+    "noble": "^1.1.0",
+    "node-watch": "^0.3.4",
+    "protocol-buffers": "^3.0.0",
+    "split": "^0.3.3",
+    "through": "^2.3.7"
+  },
+  "devDependencies": {
+    "chai": "^2.3.0",
+    "chai-as-promised": "^5.0.0",
+    "mocha": "^2.2.5",
+    "eslint": "^0.21.0"
+  }
 }
