+++ src/bluetooth/initBluetooth.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo -e 'power on\nexit' | bluetoothctl
