--- src/codec/MIN_DATE_UNIX_TIMESTAMP.js
@@ -1,5 +1,0 @@
-"use strict";
-
-var moment = require('moment');
-
-module.exports = moment("2015-05-15").unix();
