--- src/codec/MIN_SIGNAL_STRENGTH
@@ -1,3 +1,0 @@
-"use strict";
-
-module.exports = -110;
