--- src/codec/delta-encoded-measurements.proto
@@ -1,9 +1,0 @@
-message AffluenceSensorMeasurement {
-  required uint32 date = 1;
-  required bytes signal_strengths = 2 [packed=true];
-  repeated uint32 IDs = 3 [packed=true];
-}
-
-message AffluenceSensorMeasurements {
-  repeated AffluenceSensorMeasurement measurements = 1 [packed=true];
-}
