--- src/codec/measurements.proto
@@ -1,8 +1,0 @@
-message AffluenceSensorMeasurement {
-  required uint32 date = 1;
-  repeated uint32 signal_strengths = 2 [packed=true];
-}
-
-message AffluenceSensorMeasurements {
-  repeated AffluenceSensorMeasurement measurements = 1 [packed=true];
-}
