+++ src/general/hashMacAddress.js
