+++ src/wifi/formatDeviceOutput.js
