+++ src/wifi/parseOutput.js
